Expose isSerializationFirstNode and SERIALIZATION_FIRST_NODE_STRING
This exposes a mechanicsm that can be used in all the places in ember.js
/ fastboot and anywhere it might be necessary to determine whether or
not a given node is the first serialization node.  Which is useful to
ensure that the actual format can change without affecting the other
libaries who will be able to use this instead of a magic string.

The rationale for this PR can be found in greater detail at the issue
below:
https://github.com/glimmerjs/glimmer-vm/issues/787