Tack serialization-first-node-helpers onto Ember global
This is dependent on this glimmer-vm PR to actually work https://github.com/glimmerjs/glimmer-vm/pull/788. Which means we'll need to wait until it is merged/bumped and Ember itself updates to use the release with it in it before we can land this.

The idea here is to insulate things that depend on the magic string we use to determine whether or not the DOM node we pass was produced by the serialization element builder.

This needs to be exposed on the global to ensure we can use it in the ember-cli-fastboot vendor file so we can determine how to boot the app when it is loaded in the browser.

To see where it'd be used in ember-cli-fastboot please see this PR with the in progress work:

https://github.com/ember-fastboot/ember-cli-fastboot/pull/580