Add exception for `didRemoveListener` so evented proxy objects can function
When `off` is called on an Evented proxyish object it checks to see if
didRemoveListener is a function (so that it can later call it).  This
check triggers the assertion for proxy-ish objects to use `get` rather
than installed es5-getter so we add a special case to allow it.

Prior to this change when you called `off` on the proxyish object it
would trigger an assertion telling you to call
`eventedProxyObj.get('didRemoveListener') rather than
`eventedProxyObj.didRemoveListener`.

A few notes about this test:

- the unknownProperty is how we let ember know this is a proxy
- the unknownProperty function must return a value other than undefined
- Evented Mixin must be mixed in