Fix linting issues with ember-metal/lib/tracked.ts.
This file was the source of `ember-metal/lib/tracked.js` and **both**
the TS and JS versions were checked in to the repo.

The changes to tsconfig.json (to avoid shifting `lib/` out of paths, and
to include all of `packages/**/*.ts`) mean that this `tracked.ts` now
gets linted.