Ensure TestClass instance is released after test.
When create is expected to return a new instance and weakly associates the
factory manager with it.

If we return something from module scope it will hold onto the container
from module scope.

Create is expected to return a new instance and weakly associates the
factory manager with it.

If we return something from module scope it will hold onto the container
from module scope.