Begin building the Ember global dynamically.
This is another (in a long line) of interim steps to make the Ember
global auto-generated. Ultimately, we want to avoid encoding the data
structure at runtime, and instead just emit the correct
`Object.defineProperty` calls.

This interim step however, gives us a nice little win: we now lazily
require the modules used for the `Ember` global. This means we finally
stop eagerly evaluating the _entire_ `ember.prod.js` asset...