[BUGFIX release] Ensure proxies have access to `getOwner(this)`.
Prior to this change, when in debug builds, any "proxy" (i.e. something
with `unknownProperty`) would _not_ have access to an `owner` object
via `getOwner(this)`. This is because of our association with the
`FACTORY_FOR` weakmap is only setup for the actual `CoreObject` instance
being created, but when `unknownProperty` exists we never expose the
"real" `CoreObject` instance and instead expose a native JS `Proxy` (so
that we can trap usage of `instance.foo` and properly error).

This fix ensures that the native proxy that is returned is also
associated with the factory, and can properly access `getOwner(this)` /
`._debugContainerKey` / etc.

(cherry picked from commit 645fbe9378dbd95573f981ca252237a5040124b5)