[bugfix] Fix `proto` return value for native classes
Native classes end up calling the nearest class constructor, and with
the recent refactor those class constructors return their own prototype
instead of the class instance's. This breaks certain functions that
currently rely on the `proto` function return the prototype of the
instance itself. We're seeing failures in `@ember-decorators/data`
due to this bug.

This reverts to returning the protoype on the instance directly.