[bugfix] Adds options checking ability to debug/deprecation test helpers
With Ember v3 we now require that deprecations receive `id` and `until`
options. It's a hard error if they are not included, as happened
recently. This PR adds the ability to check for options when asserting
in debug helpers, so we can assert that all deprecations internally
receive these options. It also updates all outstanding cases where an
id/until value were not passed.