fix 16505: construct valid curly component manager class ref
This fixes 16505. The curly component manager was looking at the current value of its class argument reference during creation, which is never the right thing to do -- the references need to be agnostic to value because value is going to change through the lifetime of the component.

I'm using `SimpleClassNameBindingReference` here instead of `ClassNameBinding.install` because `ClassNameBinding.install` brings along more microsyntax, and AFAIK we don't support the `classNameBindings` microsyntax in template-inline class arguments like `{{some-component class=something}}`.