Fix astEqual assertion
The astEqual assertion was silently not really asserting deep equality, because normalizeNode was normalizing everything into

```
{ type: "type", body: "body" }
```

etc.

This PR restores complete checking, and miraculously no tests were failing despite this assertion not really running for quite a while.

I noticed this bug because I tried to write a failing test that refused to fail.