[BUGFIX] Fix void element printing
This PR fixes an issue where output from the printer was not actually valid HTML.
For instance preprocessing `<br>` and then printing would emit `<br></br>`, upon
parsing this the template compiler would emit an "invalid end tag" error.

This PR fixes this by propigating the `selfClosing` property from the tokenizer
into the `ElementNode`. To make sure that the syntax builder is back compat, I've
allowed the `elementNode` builder to optionally take a `TagDescriptor`, which
encapsulates the tag name and if it was authored as `selfClosing`.