Fix rehydration test suite
Currently, the rehydration-specific SSR test suite re-runs the entire SSR test suite. This is inappropriate because many of those tests do not annotate blocks in their output, so test output serialized with the rehydration builder contain additional block annotation comments that cause the tests to fail.

This commit changes the test suite so that only rehydration-specific tests run with the rehydration serializer. It also fixes a test that was failing due to a missing block annotation.