Avoid didCreateElement for EmberishGlimmerComponents.
This commit affects only the test harness, and does not in and of itself
constitute a breaking change (though the next commit will ;) ).

It is very unlikely that "glimmer components" in Ember will include
Ember.Component's legacy of always setting `id` and `class` of
`ember-view` (and in fact glimmer.js itself does _not_ do this already).
This change brings the test harness in line with that future reality.