Move Component detection from JavascriptCompiler to TemplateCompiler.
On its own, this doesn't have much impact, however future changes will
want to start adding expressions (instead of just static tag names) to
the `Component` opcode. For that to work properly it needs to happen
before the `SymbolAllocator` is ran (which promotes shared strings to a
symbols array). This change is just setting the stage for doing that
work...