Throw a compilation error for path expressions with a trailing `.`.
In prior versions of glimmer-vm the following:

```hbs
{{#if somePath.}}
  Stuff!!
{{/if}}
```

Would actually be treated like:

```hbs
{{#if somePath this}}
  Stuff!!
{{/if}}
```

This is due to how Handlebars internally parses `somePath.` (it sees two
path expressions, one for `somePath` and another for `.`).

In more recent versions of glimmer-vm we have began to throw an error
in later parts of the compilation pipeline (the `SymbolAllocator`)
because it is only able to process `PathExpressions` that have `this`
set or have contents in `parts`.

This change adds a much more helpful error.