Introduce ModifierDefinitonState
This introduces the notion of `ModifierDefinitionState`. The purpose of this state
is to encapsulate values that are required for creation of the modifier. For instance
the `ModifierDefinitionState` can contain the factory that backs a user defined modifier.
This is not a new concept and is in fact why `ComponentDefintionState` exists.

In the future there is opportunity to unify these ideas. This is the first step.