Test for API docs coverage
This adds a new test to the code-quality suite that ensure we are exporting the expected set of items into the API docs.

We have repeatedly lost parts of the API docs without noticing, so I believe this test is worth the extra friction for people who are intentionally adding or deleting public APIs.