[refactor] Adds initialization flag and isPrototypeMeta()
Currently, `meta.isInitialized()` is used for two distinct purposes:

1. To check if the meta is for an _instance_ of `EmberObject` which is
currently being _initialized_.

2. To check if a meta is the meta for a class _prototype_.

In both cases, we want similar behavior - don't trigger event listeners,
mandatory setters, or `setUnknownProperty` in some cases. So, these two
functions were tied together historically out of convenience. This PR
separates them into two methods, `isInitialized` and `isPrototypeMeta`,
to clarify their purposes and allow us to distinguish the two states.

One quirk of `isPrototypeMeta` is that it still requires an object be
passed to it. This for the case where an object does not have it's own
meta, and is relying entirely on its parent's meta. In this case, the
parent meta is being used in the context of an instance, and should be
treated as if it were an instance meta for that particular usage.