[BUGFIX] ES6 classes on/removeListener and observes/removeObserver interop v2
This is a rework of #16874 which flattens and caches the state of event
listeners more efficiently. Rather than rebuild the result of a
`matchListeners` query each time, including deduping, we flatten the
listeners down the hierarchy of metas the first time an event match is
requested. This still defers the majority of the work early on (adding
listeners is cheap) but also prevents us from having to do the work
again later.

(cherry picked from commit b3bcc7a127fd40018b99f2c41e23578bb42f813b)