Ensure dynamic component invocations follow splattribute rules.
When the `DynamicComponent` opcode was initial introduce, it did not
properly replicate the `SetComponentAttrs` toggling system that
`Ops.Component` has. This has the unfortunate result of making dynamic
component invocations not follow the established rules for
splattributes (invocation "wins" for all attributes other than `class`
and `class` is merged).

In this context `DynamicComponent` is specifically referring to angle
bracket invocation with either a named argument or path.