[BUGFIX release] fix mouseEnter/Leave event delegation w/o jQuery
The previous implementation did not work correctly when the element of
the component listening to mouseEnter/Leave was quickly moved over,
(i.e. no mouseover event triggered on this element) and the child node
is from another component, as demonstrated in the replication in #16922.

Related to 16603
Fixes #16922

(cherry picked from commit 62643188404d1b8398cb133ac30cc0e866318a8d)