`Bounds` must be non-empty
The `Bounds` object represents a _inclusive_ range. That is, the bounds
is referring to the range in DOM _including_ the `firstNode()` and the
`lastNode()`. Therefore, it does not ever make sense for those endpoints
to be `null`. (Technically, `(parent, null, null)` is a valid way to
reprsent an _empty_ range. However, in practice we do not have a use case
for this, and every use of `null` were actual bugs.)

This is technically a breaking change, but then again, if `null` were
passed today it's likely a bug, and if the consuming side was checking for
`null` it's likely just deadcode that will never execute in practice.