tests: Remove unused and broken `printTestCounts()` functions
This function is using jQuery even though it might not be available, and it is querying a selector which also no longer seems to be used by QUnit. In addition to that it is not called from anywhere and since it's broken it doesn't seem to be used manually anymore either.