Feat: Add option to turn on `ignoreStandalone` in HBS parser
Adds new config option, `ignoreStandalone`, to `preprocess()` method.
This property is passed on to the handlebars parser (default: false).
When enabled, the handlebars parser preserves more whitespace. This,
in turn, creates ASTs that, when printed back into templates, results in
more developer-friendly code.

Also bumps @types/handlebars to latest, as this is necessary for glimmer
to recognize that handlebars.parse() does accept options.