Allow iterables to use objects as keys
Previously, the key map was a dict, so all objects that #each was
iterating over needed to be coerced into a string first.

In practice, this made it impossible to use the object's identity as the
key.

This change moves the key map to a JS Map, and allows key mapping
functions to return any value (not just a string). This also makes it
possible for a host environment to implement `@identity` naturally
(without needing to stash guids or other workarounds).