Restructure code to reduce dependence on Env
The previous (and to some degree current) code treats the Environment as
an elaborate god object that host environments are required to
implement.

This change continues the process of allowing hosts to implement a
handful of delegate methods that, in practice, are the areas of
configuration, and builds the environment (and runtime context) for you.

These changes don't break existing code that hand-crafts an environment,
but the intent is to eventually moot implementations that do so by
giving them an equally expressive but much smaller API.