Move integration tests into their own package
Previously, tests lived inside of individual packages, but depended on
test-helpers, which then cyclically depended back on other packages.

After this commit, all tests that depend on more than one package live
in integration-tests, which also has the test environment and test
suites in its `lib`.

This means that the packages now work as expected by node and ecosystem
tooling.