Ensure attrSplat consumes the value that was pushed.
Prior to the changes here, for every `AttrNode` we processed we
processed its value (handling literals, mustaches, etc). The value we
process is then consumed by whatever operation is represented (e.g.
`staticArg`, `dynamicArg`, `staticAttr`, `trustingAttr`, `dynamicAttr`).
Unfortunately, when we processed the `...attributes` `AttrNode` into an
`attrSplat` we did not also ensure to consume its value (which we don't
really care about). The result was that we had one too many literal
values on the `JavascriptCompiler`'s `values` array, meaning _all_ other
operations that use a value would be getting the _prior_ operations
value (e.g. an "off by one" issue in the values array).

The fix here is to ensure that `attrSplat` consumes the value that is
pushed to ensure things stay balanced.