[FEAT] Adds autotracked to the the Glimmer references system
This PR adds autotracking around property references in Glimmer-Ember,
and adds high level acceptance tests that verify that it's working for
both components and helpers. This does _not_ test the behavior of
autotracking around modifiers, which is still ill-defined.

Note that while eventually simplifications may be possible, even
removing nested references altogether, currently it is not due to
two-way binding. Components that use two-way binding need to be able to
update a nested property reference, which means we need to store that
information somewhere accessible. It may still be possible to do this
only in cases where it is necessary (e.g. binding to a component's
arguments) but for the time being, this a smaller and iterative
approach.