[BUGFIX beta] Update Backburner.js to 2.5.0.
The primary change in Backburner.js is to use the Promise microtask
queue primarily and falling back to MutationObserver when Promise is not
present.

The test changes here are to migrate from `window.onerror` to
`window.onunhandledrejection` hooks (due to moving the microtask
delivery from `MutationObserver` to native `Promise`). The cross
platform support for `onunhandledrejection` is a tad spotty (Chrome,
Edge, and _soon_ Firefox), so I had to guard in other browsers.