Update blueprints for angle bracket syntax
Prior to this change, the blueprints generated curly brace syntax for
the component-test.

This change will default to generating the angle bracket syntax unless
it finds the component being generated has any nesting (which is not
supported by angle bracket syntax) in which case it fallsback to the
curly bracket syntax.

Fixes #17625
Based on an original implementation by @ppcano