Fix helper blueprint for module unification
Prior to this change, the classic blueprint used a different
construction for helper files. Since module unification this has
changed.

This commit updates the blueprints to generate the new construction code
for helpers when module unification is in use and still offers the
classic construction code when not.

Fixes issue #17556