[BUGFIX] Ensure tagForProperty works on class constructors
`tagForProperty` was returning the constant tag for functions, which
meant it would return the constant tag for properties on class
constructors. If a computed property was defined on the class
constructor it would fail because of this.

(cherry picked from commit 98facb100e8c72ac96261f8b05c40908633c40d0)