[BUGFIX] Adds a warning to {{action}} for undecorated methods
This PR adds a warning to the `{{action}}` modifier and the `{{action}}`
helper which tells the user they should add the `@action` decorator to
any method which is passed to it. We're only warning for the time being
since this _is_ a supported codepath, but has been actively discouraged
in the past.