[BUGFIX beta] assert `<Input>` and `<Textarea>`
`input` and `textarea` are intended to be reserved component names, but
the assertion was added to the wrong place. The end result is that only
`(component "input")` and `(component "textarea")` were disallowed, but
not when using the angle bracket syntax `<Input />` and `<Textarea />`.

This compliments the existing tests in `contextual-components-test.js`
(L1220).