[FEATURE ember-glimmer-angle-bracket-built-ins]
Add missing dependent key

There are no failing tests for this at the moment, because we only
test with all optional features on or off. It turns out that the
tracked property feature is masking this failure. As a result, the
PR that enables this feature (#17811) without tracked property has
some failing tests which should be green again after rebasing on
top of this patch.