[BUGFIX] Tracked Props - Prevents autotracking ArrayProxy creation
This PR prevents an issue with immediate invalidation within the new
autotracking transaction. ArrayProxy currently reads from one of its
own properties, `hasArrayObservers`, and then immediately updates it
with `notifyPropertyChange`.

We avoid this by using a native descriptor instead of a computed, and
not using `get()` to access it. This way, nothing is tracked during
creation (even if it is mutated).

(cherry picked from commit 03c776b20017a99cacf6a7021d247e397ce58752)