[BUGFIX] Ensure @sort works on non-Ember.Objects.
Assuming that `this.notifyPropertyChange` is a method on the object that
the `sort` is operating on is not a safe assumption. Specifically, when
operating on a `@glimmer/component` (which is _essentially_ just a very
very basic native class) there is no `notifyPropertyChange` method and
an error was thrown.

(cherry picked from commit 00a0c0bf2e4ce04e629ead7aacfef232589877dc)