Failing test for template-only components clearing
The root cause is when the `createInstance` flag is disabled, Glimmer
omits the `DidRenderLayout` opcode, which has the unintended consequence
of skipping the `popBlock` call, causing an imbalance in the stack. This
means additional content unrelated to the component will be treated as
part of the component's bounds, causing other errors later on. It also
causes other issues such as blocks not getting finalized due to the
unbalanced push and pops.

(cherry picked from commit c0aa7179863a47dfe3842b6434dadf78a01b57f2)