[BUGFIX beta] Fix #17869 (link-to causing `hash` local variable assertions)
This doesn't really "fix" the issue, it just silent this particular
assertion basically. This is ok, because the test will fail when we
implement the "everything is a value" semantics, at which point we
can fix it by renaming the hash helper. (We could fix it now by
renaming the offending hash local variable too, but that would be
pretty difficult to do correctly.)

Closes #17869.

(cherry picked from commit a39819fc8f8dc73f2cfc86db92da09921625edc3)